[ 🔧 | #47 | Refactor ] aboutPage 리팩토링
